requirements.txt
# Pin dependancies that might cause breakage
Werkzeug==2.2.2
SQLAlchemy==2.0.0

# Build dependencies
Flask==2.1.2
Flask-SQLAlchemy==3.0.2
psycopg2==2.9.5
python-dotenv==0.21.1

# Runtime dependencies
gunicorn==20.1.0
honcho==1.1.0

# Code quality
pylint==2.15.10
flake8==6.0.0

# Testing dependencies
nose==1.3.7
pinocchio==0.4.3
factory-boy==3.2.1
coverage==7.1.0
codecov==2.1.12

# Utilities
httpie==3.2.1
